refactor: refatoração da funcionalidade de atendimento
