+++ frontend/atendimento-app/data/verificacao.json
@@ -1,6 +1,6 @@
 {
   "codigo": "1234",
   "email": "teste@hotmail.com",
-  "idProfissional": "cddc8045-795c-4403-ac61-72c603aee555",
+  "idProfissional": "11111111-1111-1111-1111-111111111111",
   "urlBase": "http://localhost:8080"
 }
