+++ frontend/atendimento-app/src/features/atendimento/components/atendimentoNovoModal.tsx
